+++ src/webhook-tasks/runners/snapshotHub/index.ts
@@ -1,1 +1,1 @@
-export * from './proposalCreated';
+export * from './proposalEventRunner';
